% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{check_if_tidyproject}
\alias{check_if_tidyproject}
\title{Check if directory is a tidyproject}
\usage{
check_if_tidyproject(directory = getwd())
}
\arguments{
\item{directory}{character. working directory (default = current working directory)}
}
\description{
Check if directory is a tidyproject
}

