% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{code_library}
\alias{code_library}
\title{Show Code Library}
\usage{
code_library(extn = "mod|R|scm")
}
\arguments{
\item{extn}{vector string of extensions to include}
}
\description{
Show Code Library
}

