% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{commit_file}
\alias{commit_file}
\title{commit individual file(s)}
\usage{
commit_file(file_name)
}
\arguments{
\item{commit_file}{character vector. File(s) to be committed}
}
\description{
Has side effect that staged changed will be updated to working tree
}
\examples{
\dontrun{
commit_file("Scripts/script1.R")
}
}
