% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{createStr}
\alias{createStr}
\title{Create a folder structure for a project}
\usage{
createStr()
}
\description{
Creates a basic folder structure for a project
scripts, markdown, data and plots
}
\examples{
createStr()
}
\keyword{folder}
\keyword{structure}
