% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{dependency_tree}
\alias{dependency_tree}
\title{Create a recursive dependency_tree of file names}
\usage{
dependency_tree(from, already.got = NULL)
}
\description{
Used by copy_script. Creates an an ordered vector of script dependencies.
}

