% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{import}
\alias{import}
\title{Import staged files into project}
\usage{
import(copy_table, overwrite = FALSE, silent = FALSE)
}
\arguments{
\item{copy_table}{data frame output from stage()}

\item{overwrite}{logical (default = FALSE)}

\item{silent}{logical (default = FALSE)}
}
\description{
Import staged files into project
}
