% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_library.R
\name{import_project}
\alias{import_project}
\title{Import a project into current project}
\usage{
import_project(proj_name)
}
\arguments{
\item{proj_name}{character vector with paths to attach to}
}
\description{
If file isn't found will look in code_library
}
