% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{instCRAN}
\alias{instCRAN}
\title{install a list of packages from CRAN}
\usage{
instCRAN(list_of_packages)
}
\arguments{
\item{list_of_packages}{a character vector}
}
\description{
installs a list of packages from CRAN
provided as a character vector
}
\examples{
instCRAN(c("tidyverse", "rJava", "ggraph", "tidytree"))
}
\keyword{CRAN}
\keyword{packages}
