% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{instGH}
\alias{instGH}
\title{install a package from GitHub}
\usage{
instGH(username, pkgname)
}
\arguments{
\item{username}{a character}

\item{pkgname}{a character}
}
\description{
install a package from GitHub
providing a user name and a name of the package
}
\examples{
instGH("bonStats", "tidytreatment")
}
\keyword{GitHub}
\keyword{package}
