% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\docType{data}
\name{is_rstudio}
\alias{is_rstudio}
\title{Logical flag for detecting if R session is on rstudio not}
\format{An object of class \code{logical} of length 1.}
\usage{
is_rstudio
}
\description{
Logical flag for detecting if R session is on rstudio not
}
\keyword{datasets}
