% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{is_tidy_project}
\alias{is_tidy_project}
\title{Test if directory is a TidyProject}
\usage{
is_tidy_project(proj_path = getwd())
}
\arguments{
\item{proj_path}{character vector indicating path to TidyProject}
}
\description{
Test if directory is a TidyProject
}

