% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{list_script_dirs}
\alias{list_script_dirs}
\title{recursively get all script dirs}
\usage{
list_script_dirs(path = ".", dir_name = getOption("scripts.dir"), depth,
  maxdepth = 10)
}
\arguments{
\item{path}{character. Script dirs are in a subdirectory n levels below}

\item{dir_name}{character (default = "Scripts"). name of scripts dirs}

\item{depth}{integer. number of levels to search}
}
\description{
recursively get all script dirs
}
