% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{load_localpackage}
\alias{load_localpackage}
\title{Load local package}
\usage{
load_localpackage(...)
}
\arguments{
\item{...}{arguments to be passed to devtools::load_all}
}
\description{
Load local package
}
