% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{locate_file}
\alias{locate_file}
\title{Locate file from search path}
\usage{
locate_file(x, search.path = c("."))
}
\arguments{
\item{x}{string for file name}

\item{search.path}{vector of strings giving search path}
}
\value{
Path of located file.  Returns error if file not found.
}
\description{
Finds first file in search.path that exists
}
\examples{
\dontrun{
locate_file("script.R",c(".","Scripts")) ## looks in current working directory, then Scripts folder
}
}

