% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{make_local_bare}
\alias{make_local_bare}
\title{create local bare repository}
\usage{
make_local_bare(proj_name = getwd())
}
\arguments{
\item{proj_name}{character vector indicating path to tidyproject}
}
\description{
create local bare repository
}

