% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create new_project}
\usage{
make_project(proj_name, remove_user_lib = FALSE)
}
\arguments{
\item{proj_name}{character string of full path to new_project}

\item{remove_user_lib}{logical (default=FALSE) if TRUE will attempt to remove 
user R package library from .libPaths()}
}
\description{
Creates directory structure.  User install tidyproject again in
}
