% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{models_dir}
\alias{models_dir}
\title{Shortcut to Models directory}
\usage{
models_dir(proj_name = getwd())
}
\arguments{
\item{proj_name}{character. Working directory (default = getwd())}
}
\description{
Shortcut to Models directory
}
