% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{new_script}
\alias{new_script}
\title{Create new R script}
\usage{
new_script(name, overwrite = FALSE)
}
\arguments{
\item{name}{character indicating name of script to create}

\item{overwrite}{logical indicating whether to overwrite existing file (default=FALSE)}
}
\description{
Create new R script
}

