% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{scripts_dir}
\alias{scripts_dir}
\title{Shortcut to Scripts directory}
\usage{
scripts_dir(proj_name = getwd())
}
\arguments{
\item{proj_name}{character. Working directory (default = getwd())}
}
\description{
Shortcut to Scripts directory
}
