% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_library.R
\name{search_field}
\alias{search_field}
\title{Search for files matching field values}
\usage{
search_field(files, field, text)
}
\arguments{
\item{files}{vector string of file names/paths}

\item{field}{character. field name (can be regex) to search for}

\item{text}{string (can be regex) to search for}
}
\description{
Search for files matching field values
}

