% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_library.R
\name{search_keyword}
\alias{search_keyword}
\title{Search for files matching key words}
\usage{
search_keyword(files, text)
}
\arguments{
\item{files}{vector string of file names/paths}

\item{text}{string (can be regex) to search for}
}
\description{
Search for files matching key words
}
