% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{search_scripts}
\alias{search_scripts}
\title{Search for string in files}
\usage{
search_scripts(files, text)
}
\arguments{
\item{files}{vector string of file names/paths}

\item{text}{string (can be regex) to search for}
}
\description{
Search for string in files
}

