% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{setup_file}
\alias{setup_file}
\title{Setup files}
\usage{
setup_file(file.name, version_control = getOption("git.exists"))
}
\arguments{
\item{file.name}{character. Indicating name of file to set up}

\item{version_control}{logical. Should file be added to version control}
}
\description{
Setup files
}
