% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject-package.R
\docType{package}
\name{tidyproject-package}
\alias{tidyproject}
\alias{tidyproject-package}
\title{tidyproject: Setup the initial file structure and packages of a project}
\description{
Setups the initial file structure and install required packages for a new project.
}
\author{
\strong{Maintainer}: Paula Andrea Martinez \email{orchid00@gmail.com} (\href{https://orcid.org/0000-0002-8990-1985}{ORCID})

}
\keyword{internal}
