% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{toggle_libs}
\alias{toggle_libs}
\title{toggle library settings}
\usage{
toggle_libs(lib = c("project", "project-user", "user", "global"))
}
\arguments{
\item{lib}{character either missing, "project","project-user","user",or "global"}
}
\description{
toggle library settings
}
