% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{validate_session}
\alias{validate_session}
\title{Validate TidyProject session}
\usage{
validate_session(fail_on_error = FALSE)
}
\arguments{
\item{fail_on_error}{logical (default=FALSE) indicating whether function should give error on a test fail}
}
\description{
Validate TidyProject session
}

