% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_script.R
\name{copy_script}
\alias{copy_script}
\title{Copy script to project directory}
\usage{
copy_script(from, to, dependencies = TRUE, stamp_copy = TRUE,
  overwrite = FALSE, comment_char = "#", alt_paths, proj_path = ".")
}
\arguments{
\item{from}{character. file name or path of file to copy}

\item{to}{character. file name file to create}

\item{dependencies}{logical. Default = TRUE. will script copy dependencies}

\item{stamp_copy}{logical. Create a commented timestamp at beginning of file}

\item{overwrite}{logical. Overwrite "to" file if exists?}

\item{comment_char}{character. Comment character}

\item{alt_paths}{character vector. paths to other candidate files to search}

\item{proj_path}{character. Default = current working directory. path to tidyproject}
}
\description{
Will search code library and copy script and dependencies into scripts directory.
Script will also be stamped with source location, time and user information
}

