% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{dependency_tree}
\alias{dependency_tree}
\title{Create a recursive dependency_tree of file names}
\usage{
dependency_tree(from, already_got = NULL)
}
\arguments{
\item{from}{character. file name or path of file to copy}

\item{already_got}{logical. Default = NULL. If true will terminate loop

Used by copy_script. Creates an an ordered vector of script dependencies.}
}
\description{
Create a recursive dependency_tree of file names
}

