% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{get_script_field}
\alias{get_script_field}
\title{Get field from a code file}
\usage{
get_script_field(file_name, field_name)
}
\arguments{
\item{file_name}{character. path to file}

\item{field_name}{character. name of field to find.}
}
\description{
Get field from a code file
}

