% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{load_localpackage}
\alias{load_localpackage}
\title{Load local package}
\usage{
load_localpackage(...)
}
\arguments{
\item{...}{arguments to be passed to devtools::load_all}
}
\description{
Equivalent to  devtools::document() followed by devtools::load_all().  This
function operate on a localpackage/R directory if it exists, otherwise it
will assume the base directory is the package directory
}
\details{
if analysis directory is a package this will default to
  devtools::document() followed by devtools::load_all()
}
