% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyProject.R
\name{make_project}
\alias{make_project}
\title{Create new_project}
\usage{
make_project(proj.name, ...)
}
\arguments{
\item{proj.name}{character string of full path to new_project}

\item{...}{arguments passed to install function (default = install.packages())}
}
\description{
Create new_project
}

