% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\name{make_project}
\alias{make_project}
\title{Create new_project}
\usage{
make_project(proj_name, project_library = TRUE)
}
\arguments{
\item{proj_name}{character string of full path to new_project}

\item{project_library}{TRUE (default) or FALSE indicating whether or not to create a project library}
}
\description{
Creates directory structure.  User install tidyproject again in
}

