% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create new_project}
\usage{
make_project(
  proj_name = ".",
  lib_strategy = c("project-user", "project", "user", "global"),
  overwrite_rprofile = FALSE
)
}
\arguments{
\item{proj_name}{character string of full path to new_project}

\item{lib_strategy}{character either missing, "project","project-user","user",or "global"}

\item{overwrite_rprofile}{logical. should project .Rprofile be overwritten (default=FALSE)}
}
\description{
Creates directory structure.  User install tidyproject again in
}
