% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyproject.R
\docType{package}
\name{tidyproject}
\alias{tidyproject}
\title{Project and code management}
\description{
tidyproject is designed to manage tidy project directories and
facilitate reproducibility and reuse of code
}
\section{Strategy}{


\itemize{
 \item Code management
 \itemize{
  \item{Project structure: Defines standardised directory structure for statistical projects}
  \item{Long term reproducibility: Local (project level) installation of packages}
  \item{Code sharing/reuse: code library + search functionality}
 }
}
}

\section{Options}{

\code{scripts.dir} = names of the 'scripts' directory in the project

\code{models.dir} = names of the 'models' directory in the project

\code{git.exists} = \code{TRUE} if git is installed (only works on unix currently)

\code{code_library_path} = character vector. paths of directories containing code files

To modify these options use the options() interface.
To modify these values permanently set options() in ~/.Rprofile or R_HOME/etc/Rprofile.site
}

\examples{
\dontrun{
make_project('~/AZDXXXX/PKAE1')         ## creates new_project
code_library()                          ## display summary of code_library
preview('template1.R')                  ## preview file in code_library
copy_script('template1.R')              ## copy_script from code_library to project

new_script('test.R')                    ## Creates empty test.R
copy_script('output.data.R')            ## copies from code_library
copy_script('../pathto/script.R')       ## copies from other location
}
}
